+++ react/Image.tsx
@@ -1,6 +1,7 @@
 import React, { FunctionComponent } from 'react'
 
 import { useItemContext } from './components/ItemContext'
+import { NoImageIcon } from './components/NoImageIcon'
 import { opaque } from './utils/opaque'
 
 const Image: FunctionComponent = () => {
@@ -9,7 +10,16 @@ const Image: FunctionComponent = () => {
   return (
     <div className={opaque(item.availability)} style={{ minWidth: '96px' }}>
       <a href={item.detailUrl}>
-        <img alt={item.name} src={item.imageUrl} width="100%" />
+        {item.imageUrl ? (
+          <img
+            className="br2"
+            alt={item.name}
+            src={item.imageUrl}
+            width="100%"
+          />
+        ) : (
+          <NoImageIcon />
+        )}
       </a>
     </div>
   )