+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Icon to replace the product image when its `imageUrl` is `null`.
+
 ## [0.8.0] - 2019-10-04
 
 ### Changed